+++ api/aiconfiguration/v1alpha1/zz_generated_groupversion_info.go
@@ -33,6 +33,8 @@ func addKnownTypesGenerated(scheme *runtime.Scheme) error {
 		&AIGatewayModelProviderList{},
 		&AIGatewayPolicy{},
 		&AIGatewayPolicyList{},
+		&AIGatewaySNI{},
+		&AIGatewaySNIList{},
 	)
 
 	metav1.AddToGroupVersion(scheme, GroupVersion)