+++ modules/manager/zz_generated_konnect_controller_setup.go
@@ -24,6 +24,7 @@ func generatedControllersForKonnectEntities(
 		newKonnectEntityController[aiconfigurationv1alpha1.AIGatewayModel](controllerFactory),
 		newKonnectEntityController[aiconfigurationv1alpha1.AIGatewayModelProvider](controllerFactory),
 		newKonnectEntityController[aiconfigurationv1alpha1.AIGatewayPolicy](controllerFactory),
+		newKonnectEntityController[aiconfigurationv1alpha1.AIGatewaySNI](controllerFactory),
 		newKonnectEntityController[configurationv1alpha1.EventGatewayBackendCluster](controllerFactory),
 		newKonnectEntityController[configurationv1alpha1.EventGatewayDataPlaneCertificate](controllerFactory),
 		newKonnectEntityController[configurationv1alpha1.EventGatewayListener](controllerFactory),